Modify testing script to pull IPSW resources from the internet on first run.
This solution is chosen over checking the assets into VCS because they are too large. Using this solution means that the project is less suitable for CI infrastucture, as large files will need to be cached on build agents then mapped into test containers.

The tests do not pass on this commit.