Use BuildManifest.plist as .ipsw signature.
This is more suitable as a signature because this file is used to get the file names for the apfs and dyld shared cache .dmg files.

Adds ios16 testing for .dmg file name extraction as this is the first version where the dyld shared cache comes as a separate dmg.